refactor(statistics): Move time series visualization to component
