+++ projects/test-statistics-app/src/lib/time-series-chart-dialog.component.html
@@ -54,7 +54,7 @@
                 <div class="container-fluid">
                     <div class="row">
                         <div class="col-12">
-                            <div class="hs-statistics-timeseries" style="min-width: 400px; min-height: 50px"></div>
+                            <hs-time-series-chart [observations]="observations"></hs-time-series-chart>
                         </div>
                     </div>
                 </div>
