+++ projects/test-statistics-app/src/lib/time-series-chart.component.html
@@ -1,0 +1,1 @@
+<div class="hs-statistics-timeseries" style="min-width: 400px; min-height: 50px"></div>
